use LIMITED_ITEMS - Fix circular import
